requirements.txt
pandas==2.2.2
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
# Optional: progress bar and logging helpers
tqdm==4.66.4
structlog==24.1.0
pyarrow==17.0.0
